requirements.txt
pyserial>=3.5
kivy[base]>=2.0.0